expose numeric bitwise on Series
