+++ polars/polars-core/Cargo.toml
@@ -75,7 +75,6 @@ list = []
 rank = []
 diff = []
 moment = []
-series_bitwise = []
 
 
 # opt-in datatypes for Series
