+++ polars/polars-core/src/chunked_array/mod.rs
@@ -36,6 +36,7 @@ pub mod temporal;
 mod trusted_len;
 pub mod upstream_traits;
 use arrow::array::Array;
+mod bitwise;
 #[cfg(feature = "dtype-categorical")]
 pub(crate) mod categorical;
 pub(crate) mod list;
